requirements.txt
# Data processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# Database - MUST be < 2.0 for Airflow compatibility
sqlalchemy==1.4.50
psycopg2-binary==2.9.9

# Visualization & Reporting
matplotlib==3.8.2
fpdf2==2.7.6

# API requests
requests==2.31.0

# Airflow providers
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-slack==8.0.0
